test: expand `claim_reward` tests
